% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopusAPI.R
\name{extractXML}
\alias{extractXML}
\title{Run a search query to the Scopus API}
\usage{
extractXML(theFile)
}
\arguments{
\item{theFile}{XML file generated by search}
}
\value{
data as df
}
\description{
Run a search query to the Scopus API
}
