% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scopus.R
\name{search_scopus}
\alias{search_scopus}
\title{Run a search query to the Scopus API}
\usage{
search_scopus(api_key, string, retMax)
}
\arguments{
\item{api_key}{API key from developers portal}

\item{string}{query text}

\item{retMax}{max records to return}
}
\value{
data as df
}
\description{
Run a search query to the Scopus API
}
