% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopusAPI.R
\name{searchByString}
\alias{searchByString}
\title{Run a search query to the Scopus API}
\usage{
searchByString(
  string,
  api_key,
  content = "complete",
  myStart = 0,
  retCount = 25,
  retMax = Inf,
  mySort = "-coverDate",
  cursor = "*",
  outfile
)
}
\arguments{
\item{string}{query text}

\item{api_key}{API key from developers portal}

\item{myStart}{record to start at}

\item{retCount}{}

\item{retMax}{max records to return}

\item{mySort}{field to sort records by}
}
\value{
data as XML file
}
\description{
Run a search query to the Scopus API
}
